Add circles inside of #page-1 .content and behind its elements (transparent circle animation)
Generates circles from left to right. Each circle is a certain size and randomly generated based on a maximum value of 400. The javascript for this responsive functionality has been optimized such as to use the most appropriate variables. Timing for the fadeIn and fadeOut animations were experimented with to identify the correct timing that corresponds with circle.remove().